requirements.txt
# API requests
requests==2.32.2

# Sentiment Analysis
transformers==4.41.2
torch==2.3.1

# Data processing
pandas==2.2.2

# Snowflake
snowflake-connector-python==3.10.0
snowflake-sqlalchemy==1.5.1
sqlalchemy==1.4.49

